Refactor: useSearchParams 관련 로직 컴포넌트화
